Reorganized pages into folder, used SPA routing hack for Github pages in 404.html and index.html
