+++ starter-frontend/src/components/Navbar.js
@@ -1,9 +1,9 @@
 export default function Navbar() {
     return (
         <>
-            <nav style={{ backgroundColor: "#CBC3E3", }} className="navbar navbar-expand-lg">
+            <nav style={{ backgroundColor: "#CBC3E3", }} className="navbar navbar-expand-lg sticky-top">
                 <div className="container-fluid">
-                    <a className="navbar-brand" href="/">Navbar</a>
+                    <a className="navbar-brand" href="/">Cool Husky Games</a>
                     <button className="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavDropdown" aria-controls="navbarNavDropdown" aria-expanded="false" aria-label="Toggle navigation">
                         <span className="navbar-toggler-icon"></span>
                     </button>
