+++ starter-frontend/src/pages/Test.js
@@ -1,0 +1,7 @@
+export default function TestPage() {
+    return (
+        <main>
+            <h1>Test Page</h1>
+        </main>
+    )
+}
